#19 chore: rename ClientInfoFragment.kt to ReadClientFragment.kt to distinguish read only and write only client fragment.
