+++ presentation/src/main/java/com/kldaji/presentation/ui/client/ReadClientFragment.kt
@@ -1,4 +1,4 @@
-package com.kldaji.presentation.ui.clientInfo
+package com.kldaji.presentation.ui.client
 
 import android.os.Bundle
 import android.util.Log
@@ -7,24 +7,24 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.navArgs
-import com.kldaji.presentation.databinding.FragmentClientInfoBinding
+import com.kldaji.presentation.databinding.FragmentReadClientBinding
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class ClientInfoFragment : Fragment() {
+class ReadClientFragment : Fragment() {
     companion object {
         private const val TAG = "ClientInfoFragment"
     }
 
-    private var _binding: FragmentClientInfoBinding? = null
+    private var _binding: FragmentReadClientBinding? = null
     private val binding get() = _binding!!
-    private val navArgs: ClientInfoFragmentArgs by navArgs()
+    private val navArgs: ReadClientFragmentArgs by navArgs()
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?,
     ): View {
-        _binding = FragmentClientInfoBinding.inflate(inflater, container, false)
+        _binding = FragmentReadClientBinding.inflate(inflater, container, false)
         return binding.root
     }
