+++ presentation/src/main/java/com/kldaji/presentation/ui/clients/ClientsFragment.kt
@@ -52,7 +52,7 @@ class ClientsFragment : Fragment() {
                     true
                 }
                 R.id.add_client -> {
-                    this.findNavController().navigate(R.id.clientInfoFragment)
+                    this.findNavController().navigate(R.id.readClientFragment)
                     true
                 }
                 R.id.setting -> {
@@ -83,7 +83,7 @@ class ClientsFragment : Fragment() {
         clientAdapter = ClientAdapter(object : ClientAdapter.ItemClickListener {
             override fun onItemClick(client: Client) {
                 val direction =
-                    ClientsFragmentDirections.actionClientsFragmentToClientInfoFragment(client)
+                    ClientsFragmentDirections.actionClientsFragmentToReadClientFragment(client)
                 this@ClientsFragment.findNavController().navigate(direction)
             }
         })
