+++ presentation/src/main/res/layout/fragment_read_client.xml
@@ -3,11 +3,11 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.kldaji.presentation.ui.clientInfo.ClientInfoFragment">
+    tools:context="com.kldaji.presentation.ui.client.ReadClientFragment">
 
     <TextView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:text="ClientInfo Fragment" />
+        android:text="ReadClient Fragment" />
 
 </FrameLayout>
