+++ presentation/src/main/res/navigation/nav_graph.xml
@@ -14,8 +14,8 @@
             android:id="@+id/action_clientsFragment_to_searchFragment"
             app:destination="@id/searchFragment" />
         <action
-            android:id="@+id/action_clientsFragment_to_clientInfoFragment"
-            app:destination="@id/clientInfoFragment" />
+            android:id="@+id/action_clientsFragment_to_readClientFragment"
+            app:destination="@id/readClientFragment" />
         <action
             android:id="@+id/action_clientsFragment_to_settingFragment"
             app:destination="@id/settingFragment" />
@@ -32,10 +32,10 @@
         android:label="fragment_search"
         tools:layout="@layout/fragment_search" />
     <fragment
-        android:id="@+id/clientInfoFragment"
-        android:name="com.kldaji.presentation.ui.clientInfo.ClientInfoFragment"
-        android:label="fragment_client_info"
-        tools:layout="@layout/fragment_client_info" >
+        android:id="@+id/readClientFragment"
+        android:name="com.kldaji.presentation.ui.client.ReadClientFragment"
+        android:label="fragment_read_client"
+        tools:layout="@layout/fragment_read_client" >
         <argument
             android:name="client"
             app:argType="com.kldaji.domain.Client"
